integration_tests/requirements.txt
pytest
pytest-docker
openai
requests
python-keycloak
PyJWT
cryptography
